Changed the draugr attacks to be decided by two separate methods depending on whether the attack is capable of poisoning or not.
Added the smokestack attacks to the enemy files.

Added the text output to tell the user they are burning when the burning debuff is active, as well as the text to report whether the player has been set on fire or become necrotic.

Added the text output to introduce the poison boss

Fixed a bug where the player could spawn in an empty room.

Added a method to check for player debuffs and inflict burning damage, player how has a health bar and a text output health bar.

Rooms now generate enemies based on a variable determined by room difficulty instead of a constant of 3.