Bosses now have lines they say before attacks.
Fixed a bug where the enemy name and attack were flipped when outputted to the user

Added Javadocs to GameIO.java, Room.java and Map.java methods that required it.

Added outputs to signal which encounter the player is on per boss fight and to signal the beginning of the phase interim battle

Changed GameIO.sleep() to take the number of milliseconds the thread should sleep for as a parameter to allow for variable wait times.