Fixed an issue in GameIO where the wrong directions were being printed for the rooms when the player was picking a room.
Removed the ability to see the room difficulty when the map is printed to the user, player now shows up as an X on the map in the room they are current occupying.

Fixed a bug where the wrong method was being called for a boss encounter, causing the boss encounter to never occur.

Changed the name of the enemies array in the room to waveOne, so it can be expanded upon in more difficult rooms.